everything appears to be working at the moment

+++ apps/api/post_classes.py
@@ -11,7 +11,7 @@ class NewUser(BaseModel):
 class IntervalWorkout(BaseModel):
     user_id:str = 'guest'
     distance:int = 0
-    time:int = 0
+    time:float = 0
     rest:int = 0
     intervals:int = 1
     date = '2000-01-01'
@@ -20,9 +20,9 @@ class IntervalWorkout(BaseModel):
 class NewWorkout(BaseModel):
     user_id:int
     workout_date:str = '2000-01-01'
-    time_sec:int = 0
+    time_sec:float = 0.0
     distance:int = 0
-    split:int = 0
+    split:float = 0.0
     sr:int = 0
     hr:int = 0
     intervals:int = 1
@@ -30,9 +30,9 @@ class NewWorkout(BaseModel):
 
 class NewInterval(BaseModel):
     workout_id:int 
-    time_sec:int
+    time_sec:float
     distance:int 
-    split:int=0
+    split:float=0.0
     sr:int=0
     hr:int=0
     rest:int=0
